pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.1"
description = ""
authors = ["Konstantin Vlasov <ke.vlasov@yandex.ru>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

repository = "https://github.com/kvlas/python-project-83"
homepage = "https://github.com/kvlas/python-project-83"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
gunicorn = "^23.0.0"
pyyaml = "^6.0.1"
jinja2 = "^3.1.4"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
responses = "^0.25.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"